refac(various) updated imports for redux methods, refactored out redux from main
