+++ src/components/redux-data/index.js
@@ -1,0 +1,4 @@
+import ReduxMainData from './ReduxMainData';
+import ReduxSingleRaceData from './ReduxSingleRaceData';
+
+export { ReduxMainData, ReduxSingleRaceData };
